requirements.txt
urllib3==1.25.10
chardet==3.0.4
requests==2.23.0
splinter
boto3==1.17.76
botocore==1.20.76
awswrangler==2.8.0
selenium==3.141.0
Pyarrow
#Pyarrow==0.17.0
snowflake-connector-python==2.3.7
